requirements.txt
torch~=1.12.0
torchvision~=0.8.2
pillow~=9.1.1